feat(linter): include `local` executor in exclusion list
This update adds the `local` executor to the list of executors to be
flagged, since it is not suitable for cloud configurations. Also update
the variable name to make more general and thus better reflect the
executors in the list.

Also added a unit test to confirm that `local` is appropriately
flagged when specified.